+++ pkg/itest/list_test.go
@@ -1,0 +1,19 @@
+//go:build integration
+// +build integration
+
+package cmd_test
+
+import (
+	"testing"
+)
+
+func TestList(t *testing.T) {
+	err := runSingle(t, nil,
+		"plan",
+		"list",
+	)
+
+	if err != nil {
+		t.Fatal(err)
+	}
+}