+++ plans/example-browser-node/src/index.js
@@ -1,0 +1,12 @@
+const { invokeMap } = require('@testground/sdk')
+
+const testcases = {
+  failure: require('./failure'),
+  output: require('./output'),
+  sync: require('./sync')
+}
+
+;(async () => {
+  // This is the plan entry point.
+  await invokeMap(testcases)
+})()