+++ CHANGELOG.md
@@ -7,11 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.0] - unreleased
 ### Added
 - This ̀`CHANGELOG.md` file. See [PR 1445]
-- Support addition of containers to the control network See [PR 1481]
+- Support addition of containers to the control network. See [PR 1481]
+- Add `pick`, `toml`, `withEnv`, and `atoi` to templates. See [PR 1516]
+- Add `[[runs]]` field to compositions. See [PR 1516]
+- Add `--run-ids` option during `testground run`. See [PR 1516]
+- Add `--result-file ./output.csv` option during `testground run`. See [PR 1516]
 
 ### Fixed
 - Fix dependencies rewrites in the `exec:go` builder. See [PR 1469]
 
+
+### Changed
+- Port shell code used for integration testing to a go package. See [PR 1537]
+
 [PR 1445]: https://github.com/testground/testground/pull/1445
 [PR 1469]: https://github.com/testground/testground/pull/1469
 [PR 1481]: https://github.com/testground/testground/pull/1481
+[PR 1537]: https://github.com/testground/testground/pull/1537